+++ acc_app_optimisation/job_control/single_objective/jobs.py
@@ -4,6 +4,7 @@ from logging import getLogger
 
 import gym
 import numpy as np
+from cernml import coi
 from cernml.coi import SingleOptimizable
 from cernml.coi.mpl_utils import iter_matplotlib_figures
 from cernml.coi_funcs import FunctionOptimizable
@@ -11,7 +12,7 @@ from PyQt5 import QtCore
 
 from ...envs import Metadata
 from ...utils.bounded import BoundedArray
-from ..base import Job, JobCancelled
+from ..base import Job
 from . import constraints, optimizers
 
 LOG = getLogger(__name__)
@@ -37,10 +38,11 @@ class OptJob(Job):
     def __init__(
         self,
         *,
+        token_source: coi.CancellationTokenSource,
         signals: Signals,
         problem: optimizers.Optimizable,
     ) -> None:
-        super().__init__()
+        super().__init__(token_source)
         self.problem = problem
         self.wrapped_constraints = [
             constraints.CachedNonlinearConstraint.from_any_constraint(c)
@@ -73,7 +75,7 @@ class OptJob(Job):
         # pylint: disable = bare-except
         try:
             self.run_optimization()
-        except JobCancelled:
+        except coi.CancelledError:
             LOG.info("Optimization cancelled")
         except:
             LOG.error(traceback.format_exc())
@@ -84,7 +86,7 @@ class OptJob(Job):
 
     def _env_callback(self, action: np.ndarray) -> float:
         """The callback function provided to BaseOptimizer.solve()."""
-        self.cancellation_token.raise_if_cancelled()
+        self._token_source.token.raise_if_cancellation_requested()
         # Yield at least once per optimization step. This releases
         # Python's Global Interpreter Lock (GIL) and gives the main
         # thread a chance to process GUI events.
@@ -149,11 +151,12 @@ class SingleOptimizableJob(OptJob):
     def __init__(
         self,
         *,
+        token_source: coi.CancellationTokenSource,
         signals: Signals,
         problem: SingleOptimizable,
         optimizer_factory: optimizers.OptimizerFactory,
     ) -> None:
-        super().__init__(signals=signals, problem=problem)
+        super().__init__(token_source=token_source, signals=signals, problem=problem)
         self.x_0 = self.problem.get_initial_params()
         bounds = (problem.optimization_space.low, problem.optimization_space.high)
         self._solve = optimizer_factory.make_solve_func(
@@ -182,12 +185,13 @@ class FunctionOptimizableJob(OptJob):
     def __init__(
         self,
         *,
+        token_source: coi.CancellationTokenSource,
         signals: Signals,
         problem: FunctionOptimizable,
         optimizer_factory: optimizers.OptimizerFactory,
         skeleton_points: t.Iterable[float],
     ) -> None:
-        super().__init__(signals=signals, problem=problem)
+        super().__init__(token_source=token_source, signals=signals, problem=problem)
         self.skeleton_points = tuple(skeleton_points)
         self.all_x_0 = [
             problem.get_initial_params(point) for point in self.skeleton_points